custom-mcp-servers/document-generator-mcp/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-docx==0.8.11
numpy==1.24.3
pandas==2.0.3
python-multipart==0.0.6
